// packages/database/migrations/snapshots/22f63f0bf7de112e0d669b09ab52a93e6d2ca6ccf851724fc9017bd61125bf28/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Enemy: {
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly hp: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly locationId: CodecTypes['pg/text@1']['output'];
      readonly maxHp: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Game: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly GameMessage: {
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly gameId: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly role: CodecTypes['pg/text@1']['output'];
    };
    readonly InventoryItem: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly itemId: CodecTypes['pg/text@1']['output'];
      readonly playerId: CodecTypes['pg/text@1']['output'];
      readonly quantity: CodecTypes['pg/int4@1']['output'];
    };
    readonly Item: {
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly locationId: CodecTypes['pg/text@1']['output'] | null;
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Location: {
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly gameId: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly LocationConnection: {
      readonly fromLocationId: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly toLocationId: CodecTypes['pg/text@1']['output'];
    };
    readonly NPC: {
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly locationId: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Player: {
      readonly gameId: CodecTypes['pg/text@1']['output'];
      readonly gold: CodecTypes['pg/int4@1']['output'];
      readonly hp: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly level: CodecTypes['pg/int4@1']['output'];
      readonly locationId: CodecTypes['pg/text@1']['output'];
      readonly maxHp: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Enemy: {
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly hp: CodecTypes['pg/int4@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly locationId: CodecTypes['pg/text@1']['input'];
      readonly maxHp: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Game: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly GameMessage: {
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly gameId: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly role: CodecTypes['pg/text@1']['input'];
    };
    readonly InventoryItem: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly itemId: CodecTypes['pg/text@1']['input'];
      readonly playerId: CodecTypes['pg/text@1']['input'];
      readonly quantity: CodecTypes['pg/int4@1']['input'];
    };
    readonly Item: {
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly locationId: CodecTypes['pg/text@1']['input'] | null;
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Location: {
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly gameId: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly LocationConnection: {
      readonly fromLocationId: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly toLocationId: CodecTypes['pg/text@1']['input'];
    };
    readonly NPC: {
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly locationId: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Player: {
      readonly gameId: CodecTypes['pg/text@1']['input'];
      readonly gold: CodecTypes['pg/int4@1']['input'];
      readonly hp: CodecTypes['pg/int4@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly level: CodecTypes['pg/int4@1']['input'];
      readonly locationId: CodecTypes['pg/text@1']['input'];
      readonly maxHp: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly Enemy: {
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly hp: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly locationId: CodecTypes['pg/text@1']['output'];
      readonly maxHp: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Game: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly GameMessage: {
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly gameId: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly role: CodecTypes['pg/text@1']['output'];
    };
    readonly InventoryItem: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly itemId: CodecTypes['pg/text@1']['output'];
      readonly playerId: CodecTypes['pg/text@1']['output'];
      readonly quantity: CodecTypes['pg/int4@1']['output'];
    };
    readonly Item: {
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly locationId: CodecTypes['pg/text@1']['output'] | null;
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Location: {
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly gameId: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly LocationConnection: {
      readonly fromLocationId: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly toLocationId: CodecTypes['pg/text@1']['output'];
    };
    readonly NPC: {
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly locationId: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Player: {
      readonly gameId: CodecTypes['pg/text@1']['output'];
      readonly gold: CodecTypes['pg/int4@1']['output'];
      readonly hp: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly level: CodecTypes['pg/int4@1']['output'];
      readonly locationId: CodecTypes['pg/text@1']['output'];
      readonly maxHp: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly Enemy: {
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly hp: CodecTypes['pg/int4@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly locationId: CodecTypes['pg/text@1']['input'];
      readonly maxHp: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Game: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly GameMessage: {
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly gameId: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly role: CodecTypes['pg/text@1']['input'];
    };
    readonly InventoryItem: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly itemId: CodecTypes['pg/text@1']['input'];
      readonly playerId: CodecTypes['pg/text@1']['input'];
      readonly quantity: CodecTypes['pg/int4@1']['input'];
    };
    readonly Item: {
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly locationId: CodecTypes['pg/text@1']['input'] | null;
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Location: {
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly gameId: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly LocationConnection: {
      readonly fromLocationId: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly toLocationId: CodecTypes['pg/text@1']['input'];
    };
    readonly NPC: {
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly locationId: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Player: {
      readonly gameId: CodecTypes['pg/text@1']['input'];
      readonly gold: CodecTypes['pg/int4@1']['input'];
      readonly hp: CodecTypes['pg/int4@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly level: CodecTypes['pg/int4@1']['input'];
      readonly locationId: CodecTypes['pg/text@1']['input'];
      readonly maxHp: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;